VIS-1224: Explorer chart pane drops its insight section; Add Insight = create-only
In the Explorer the chart's insights ARE the stacked insight panes above it, so
the chart pane no longer carries its own insight list / drop zone / add menu:

- ChartBuildSection: removed the insight-selection section entirely — the chart
  pane is now just Basic Information + Layout (the standard chart chrome).
- ExplorationBuildRail: "+ Add Insight" (below the stack) is now a plain button
  that appends a new blank insight — you create insights in the Explorer, you
  don't pick existing ones there.
- AddInsightMenu (the New-blank / Find-existing dropdown) extracted to a shared
  component (workspace/AddInsightMenu.jsx), parameterized with `excludeNames`,
  ready for the standard chart edit panel to use next.
- Tests migrated (rail Add Insight = create; chart-pane insight/drop-zone tests
  removed); 4 e2e stories drop the now-dead create-new menu step. Full
  workspace/explorer/common sweep: 4521 pass, lint clean.
