+++ viewer/e2e/stories/save-as-metric.spec.mjs
@@ -272,12 +272,6 @@ test.describe('Save as metric (Explore 2.0 Phase 4 — 06 §4/§8)', () => {
     );
 
     await page.getByTestId('right-panel-add-insight').click();
-    // Phase 6c-T5 (ux-audit.md "'+ Add Insight' creates a blank insight instead
-    // of letting you pick an existing one"): the button now opens a picker
-    // (existing insights + "New blank insight"); these specs want the OLD
-    // "always create a fresh blank insight" behavior, so drive the new
-    // secondary action explicitly.
-    await page.getByTestId('add-insight-menu-create-new').click();
     const insightNames = await page.evaluate(
       () => window.useStore.getState().explorerChartInsightNames
     );
@@ -347,12 +341,6 @@ test.describe('Save as metric (Explore 2.0 Phase 4 — 06 §4/§8)', () => {
     );
 
     await page.getByTestId('right-panel-add-insight').click();
-    // Phase 6c-T5 (ux-audit.md "'+ Add Insight' creates a blank insight instead
-    // of letting you pick an existing one"): the button now opens a picker
-    // (existing insights + "New blank insight"); these specs want the OLD
-    // "always create a fresh blank insight" behavior, so drive the new
-    // secondary action explicitly.
-    await page.getByTestId('add-insight-menu-create-new').click();
     const insightNames = await page.evaluate(
       () => window.useStore.getState().explorerChartInsightNames
     );
