+++ viewer/src/components/views/workspace/ChartBuildSection.jsx
@@ -1,51 +1,25 @@
 import React, { useState, useEffect, useCallback, useRef } from 'react';
-import { PiCaretDown, PiCaretRight, PiPlus, PiX, PiPencilSimple } from 'react-icons/pi';
-import { useDroppable } from '@dnd-kit/core';
-import EmbeddedPill from '../lineage/EmbeddedPill';
+import { PiCaretDown, PiCaretRight, PiX, PiPencilSimple } from 'react-icons/pi';
 import useStore from '../../../stores/store';
 import PanelMenu from '../../common/PanelMenu';
-import { selectInsightStatus } from '../../../stores/explorerStore';
 import { getSchema } from '../../../schemas/schemas';
 import ChartEditFormFields from '../common/ChartEditFormFields';
-import { recordOnboardingAction } from '../../onboarding/onboardingState';
-
-const InsightPillItem = ({ name, isActive, onRemove, onClick }) => {
-  const status = useStore(selectInsightStatus(name));
-  return (
-    <span data-testid={`chart-insight-pill-${name}`}>
-      <EmbeddedPill
-        objectType="insight"
-        label={name}
-        isActive={isActive}
-        onClick={onClick}
-        onRemove={onRemove}
-        statusDot={status}
-      />
-    </span>
-  );
-};
 
 /**
- * ChartBuildSection — the Explorer Build-rail chart pane. VIS-1224: the colored
- * side-bar body is gone; it now renders the SAME standard chart edit panel as
- * the RightRail (`ChartEditFormFields` — Basic Information + Layout) with the
- * Explorer's own insight-selection section (drop zone + activate-on-click pills
- * + Add Insight) passed in as the shared panel's `insightsSection` slot. Name
- * edits write through `setChartName` live (no Save button); the ⋮ menu's
- * "Rename" focuses the Basic Information name field (disabled for a loaded/saved
- * chart — that rename is VIS-1209's project-wide ${ref()} rewrite).
+ * ChartBuildSection — the Explorer Build-rail chart pane. VIS-1224: renders the
+ * SAME standard chart edit panel as the RightRail (`ChartEditFormFields` —
+ * Basic Information + Layout). It has NO insight-selection section: in the
+ * Explorer the chart's insights ARE the stacked insight panes above it (added
+ * via the rail's "+ Add Insight"), so there is nothing to pick or drop here.
+ * Name edits write through `setChartName` live (editable only for an unsaved
+ * chart; the ⋮ Rename focuses the Basic Information field).
  */
 const ChartBuildSection = ({ isExpanded, onToggleExpand }) => {
   const isLoadedChart = useStore(s => (s.charts || []).some(c => c.name === s.explorerChartName));
   const chartName = useStore(s => s.explorerChartName);
   const chartLayout = useStore(s => s.explorerChartLayout);
-  const chartInsightNames = useStore(s => s.explorerChartInsightNames);
-  const activeInsightName = useStore(s => s.explorerActiveInsightName);
   const setChartName = useStore(s => s.setChartName);
   const replaceChartLayout = useStore(s => s.replaceChartLayout);
-  const createInsight = useStore(s => s.createInsight);
-  const removeInsightFromChart = useStore(s => s.removeInsightFromChart);
-  const setActiveInsight = useStore(s => s.setActiveInsight);
   const closeChart = useStore(s => s.closeChart);
 
   const [layoutSchema, setLayoutSchema] = useState(null);
@@ -63,11 +37,6 @@ const ChartBuildSection = ({ isExpanded, onToggleExpand }) => {
       setRenameError(null);
     }
   }, [chartName, isEditing]);
-
-  const { setNodeRef: setInsightDropRef, isOver: isInsightOver } = useDroppable({
-    id: 'chart-insight-zone',
-    data: { type: 'insight-zone' },
-  });
 
   useEffect(() => {
     let cancelled = false;
@@ -114,26 +83,6 @@ const ChartBuildSection = ({ isExpanded, onToggleExpand }) => {
     [replaceChartLayout]
   );
 
-  const handleAddInsight = useCallback(() => {
-    createInsight();
-    recordOnboardingAction('insight_added');
-  }, [createInsight]);
-
-  const handleRemoveInsight = useCallback(
-    (e, name) => {
-      e.stopPropagation();
-      removeInsightFromChart(name);
-    },
-    [removeInsightFromChart]
-  );
-
-  const handleInsightClick = useCallback(
-    name => {
-      setActiveInsight(name);
-    },
-    [setActiveInsight]
-  );
-
   const handleToggle = useCallback(
     e => {
       e.stopPropagation();
@@ -148,46 +97,6 @@ const ChartBuildSection = ({ isExpanded, onToggleExpand }) => {
       closeChart?.();
     },
     [closeChart]
-  );
-
-  // The Explorer-specific insight selection — a drop zone (chart-insight-zone),
-  // activate-on-click pills, and Add Insight — handed to the shared panel as
-  // its insight-selection slot.
-  const insightsSection = (
-    <div
-      ref={setInsightDropRef}
-      data-testid="chart-insight-drop-zone"
-      className={`space-y-1 rounded p-2 transition-all ${
-        isInsightOver ? 'ring-2 ring-primary-400 ring-offset-1 bg-primary-50/50' : ''
-      }`}
-    >
-      <h3 className="text-sm font-medium text-gray-700 border-b border-gray-200 pb-2">Insights</h3>
-      {chartInsightNames.length === 0 ? (
-        <p className="text-xs text-gray-400 py-2">
-          No insights added yet. Drag from the Library or click below.
-        </p>
-      ) : (
-        <div className="flex flex-wrap gap-1.5">
-          {chartInsightNames.map(name => (
-            <InsightPillItem
-              key={name}
-              name={name}
-              isActive={name === activeInsightName}
-              onClick={() => handleInsightClick(name)}
-              onRemove={e => handleRemoveInsight(e, name)}
-            />
-          ))}
-        </div>
-      )}
-      <button
-        data-testid="chart-add-insight"
-        onClick={handleAddInsight}
-        className="flex items-center gap-1 mt-2 text-xs text-primary-600 hover:text-primary-800 transition-colors"
-      >
-        <PiPlus size={12} />
-        Add Insight
-      </button>
-    </div>
   );
 
   return (
@@ -268,7 +177,6 @@ const ChartBuildSection = ({ isExpanded, onToggleExpand }) => {
             nameErrorTestId="chart-rename-error"
             nameInputRef={nameInputRef}
             nameTestId="chart-name-input"
-            insightsSection={insightsSection}
             layoutTitle="Layout Properties"
             layoutSchema={layoutSchema}
             layoutValues={chartLayout}
